Reset compiler version with new warnings to 12.0.5.
Reverts commit 313fe8. It seems that the new warnings are due to a
a compiler option being set by Xcode 12.5 rather than the new version
of the compiler. However CMake provides no insight into the Xcode
version so revert to not trying to turn off the warnings for version
12.0.0 to avoid the unknown command option warnings in CI which
currently uses Xcode 12.4 and clang 12.0.0. We'll tolerate the
compile warnings, that we are trying to disable, in local builds
in the event that CMake/Xcode 12.5 somehow selects clang 12.0.0
instead of 12.0.5.